requirements_dev.txt
alabaster==0.7.12
argh==0.26.2
Babel==2.9.1
backports.entry-points-selectable==1.1.0
bleach==4.0.0
bump2version==0.5.11
certifi==2021.5.30
charset-normalizer==2.0.4
Click==7.0
coverage==4.5.4
distlib==0.3.2
docutils==0.17.1
entrypoints==0.3
filelock==3.0.12
flake8==3.7.8
idna==3.2
imagesize==1.2.0
Jinja2==3.0.1
MarkupSafe==2.0.1
mccabe==0.6.1
packaging==21.0
pathtools==0.1.2
Pillow==8.3.1
pkginfo==1.7.1
platformdirs==2.2.0
pluggy==0.13.1
py==1.10.0
pycodestyle==2.5.0
pyflakes==2.1.1
pygame==2.0.1
Pygments==2.10.0
pyparsing==2.4.7
pytz==2021.1
PyYAML==5.4.1
readme-renderer==29.0
requests==2.26.0
requests-toolbelt==0.9.1
six==1.16.0
snowballstemmer==2.1.0
Sphinx==1.8.5
sphinxcontrib-serializinghtml==1.1.5
sphinxcontrib-websupport==1.2.4
toml==0.10.2
tox==3.14.0
tqdm==4.62.1
twine==1.14.0
urllib3==1.26.6
virtualenv==20.7.2
watchdog==0.9.0
webencodings==0.5.1

setuptools~=44.1.1
numpy~=1.21.2
